Script is ready but won't be merged. Waiting clarification for common functions.
